feature: update dependencies
