+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 <!-- markdownlint-disable-file MD004 MD024 MD034 MD036 -->
 # CHANGE LOG
 
+## main(v1.0.1)
+
 ## v1.0.0
 
 - fix: |UI| 修复 User 查看收件箱，不选择地址时，关键词查询不生效
