+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cloudflare_temp_email",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "private": true,
   "type": "module",
   "scripts": {
@@ -32,7 +32,7 @@
     "vooks": "^0.2.12",
     "vue": "^3.5.17",
     "vue-clipboard3": "^2.0.0",
-    "vue-i18n": "^11.1.6",
+    "vue-i18n": "^11.1.7",
     "vue-router": "^4.5.1"
   },
   "devDependencies": {
@@ -47,7 +47,7 @@
     "vite-plugin-wasm": "^3.4.1",
     "workbox-build": "^7.3.0",
     "workbox-window": "^7.3.0",
-    "wrangler": "^4.20.4"
+    "wrangler": "^4.21.0"
   },
   "packageManager": "pnpm@10.10.0+sha512.d615db246fe70f25dcfea6d8d73dee782ce23e2245e3c4f6f888249fb568149318637dca73c2c5c8ef2a4ca0d5657fb9567188bfab47f566d1ee6ce987815c39"
 }
