+++ worker/src/constants.ts
@@ -1,5 +1,5 @@
 export const CONSTANTS = {
-    VERSION: 'v' + '1.0.0',
+    VERSION: 'v' + '1.0.1',
 
     // DB Version
     DB_VERSION_KEY: 'db_version',
